Optimization - Move email check to a middleware
