+++ routes/users.js
@@ -3,10 +3,11 @@ const postUsers = require('../controllers/postUsers')
 const {
   get
 } = require('../controllers/users')
+const checkEmail = require('../middlewares/checkEmail')
 
 const router = express.Router()
 
 router.get('/', get)
-router.post('/', postUsers)
+router.post('/', checkEmail, postUsers)
 
 module.exports = router
